requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
langchain
langserve
sse_starlette
httpx<0.27
pydantic==1.10.13